+++ libraries/bio/src/pdb/auto-dock-service.ts
@@ -54,15 +54,20 @@ export interface IAutoDockService {
 
   startDockerContainer(timeout?: number): Promise<void>;
 
+  checkOpenCl(): Promise<number>;
+
   /**
    * @param receptor  PDB string of target (receptor) structure
    * @param ligand    PDB or mol structure of the ligand to be docked
    * @param npts      Grid
    */
-  run(
-    receptor: string, ligand: BiostructureData, npts: GridSize, poseCount?: number,
+  dockLigand(receptor: BiostructureData, ligand: BiostructureData, autodockGpf: string, poseCount?: number,
     posColName?: string
-  ): Promise<AutoDockRunResult>;
+  ): Promise<DG.DataFrame>;
+
+  dockLigandColumn(receptor: BiostructureData, ligandCol: DG.Column<string>, autodockGpf: string, poseCount?: number,
+    posColName?: string
+  ): Promise<DG.DataFrame>;
 }
 
 export async function getAutoDockService(): Promise<IAutoDockService> {