+++ packages/Docking/src/utils/auto-dock-service.ts
@@ -9,26 +9,41 @@ import {
 import {BiostructureData, BiostructureDataJson} from '@datagrok-libraries/bio/src/pdb/types';
 import {getPdbHelper, IPdbHelper} from '@datagrok-libraries/bio/src/pdb/pdb-helper';
 import {DockerContainerStatus, awaitStatus} from '@datagrok-libraries/bio/src/utils/docker';
-
+import {delay, expectExceptionAsync} from '@datagrok-libraries/utils/src/test';
+import {Molecule3DUnitsHandler} from '@datagrok-libraries/bio/src/molecule-3d/molecule-3d-units-handler';
+import {MoleculeUnitsHandler} from '@datagrok-libraries/bio/src/molecule/molecule-units-handler';
 
 import {_package, getAutoDockService} from '../package';
-import {DockerContainer} from 'datagrok-api/dg';
-import {delay} from '@datagrok-libraries/utils/src/test';
 
 namespace Forms {
-  export type run = {
+  export type dockLigand = {
     /** PDB string of the receptor target structure */ receptor: string,
+    /** */ receptor_format: string,
     /** PDB string of the ligand structure */ ligand: string, /* TODO: BiostructureDataJson support */
-    /** */ poseCount?: number,
+    /** */ ligand_format: string,
+    /** */ autodock_gpf: string,
+    /** */ pose_count: number,
   }
-  export type runRes = {
+  export type dockLigandRes = {
     /** pdbqt */ poses: string;
+  }
+
+  export type dockLigandList = {
+    receptor: string,
+    receptor_format: string,
+    ligand: string[],
+    ligand_format: string,
+    autodock_gpf: string,
+    pose_count: number,
+  }
+  export type dockLigandListRes = {
+    /** pdbqt */ ligand_results: { [ligandIdx: number]: string; }
   }
 }
 
-export function buildDefaultGridConfig(receptorName: string, gridSize: GridSize): string {
+export function buildDefaultAutodockGpf(receptorName: string, npts: GridSize): string {
   const res: string = `
-npts ${gridSize.x} ${gridSize.y} ${gridSize.y}      # num.grid points in xyz
+npts ${npts.x} ${npts.y} ${npts.y}      # num.grid points in xyz
 gridfld ${receptorName}.maps.fld                    # grid_data_file
 spacing 0.375                                       # spacing(A)
 receptor_types A C Fe N NA OA SA                    # receptor atom types
@@ -110,18 +125,41 @@ export class AutoDockService implements IAutoDockService {
     return awaitStatus(this.dc.id, targetStatus, timeout, _package.logger);
   }
 
-  async run(
-    receptor: string, ligand: BiostructureData, npts: GridSize, poseCount?: number,
-    poseColName?: string, debug: boolean = false
-  ): Promise<AutoDockRunResult> {
+  // -- Methods --
+
+  async checkOpenCl(): Promise<number> {
+    const path = '/check_opencl';
+    const params: RequestInit = {
+      method: 'GET',
+    };
+    const adRes = (await this.fetchAndCheck(path, params));
+    const clinfoSuccess: boolean = adRes['success'];
+    if (!clinfoSuccess)
+      throw new Error(adRes['error']);
+
+    const clinfoOut: string = adRes['output'];
+    const clinfoOutMa = clinfoOut.match(/.+platform.+\s+(?<count>\d)/);
+    if (!clinfoOutMa)
+      throw new Error('Unexpected clinfo output');
+    const clinfoCount = parseInt(clinfoOutMa.groups!['count']);
+    return clinfoCount;
+  }
+
+  async dockLigand(receptor: BiostructureData, ligand: BiostructureData,
+    autodockGpf: string, poseCount: number = 30, poseColName: string = 'poses', debug: boolean = false
+  ): Promise<DG.DataFrame> {
+    if (receptor.binary || receptor.ext !== 'pdb')
+      throw new Error(`Unsupported receptor ext '${receptor.ext}' or binary, must be 'pdb' string.`);
     if (ligand.binary || ligand.ext !== 'pdb')
       throw new Error(`Unsupported ligand ext '${ligand.ext}' or binary, must be 'pdb' string.`);
-    const ligandPdb = ligand.data as string;
 
-    const form: Forms.run = {
-      receptor: receptor,
-      ligand: ligandPdb,
-      poseCount: poseCount,
+    const form: Forms.dockLigand = {
+      receptor: receptor.data as string,
+      receptor_format: receptor.ext,
+      ligand: ligand.data as string,
+      ligand_format: ligand.ext,
+      autodock_gpf: autodockGpf,
+      pose_count: poseCount,
     };
     const params: RequestInit = {
       method: 'POST',
@@ -129,16 +167,11 @@ export class AutoDockService implements IAutoDockService {
       body: JSON.stringify(form),
     };
 
-    const path = `/dock?x=${npts.x}&y=${npts.y}&z=${npts.z}&debug=${debug}`;
-    const adResStr = (await grok.dapi.docker.dockerContainers.request(this.dc.id, path, params))!;
+    const path = `/autodock/dock_ligand`;
+    // const adResStr = (await grok.dapi.docker.dockerContainers.request(this.dc.id, path, params))!;
+    // const adRes: Forms.runRes = JSON.parse(adResStr) as Forms.runRes;
     // TODO: Use the new dockerContainers API
-    // const adResponse: Response = await grok.dapi.docker.dockerContainers.fetchProxy(this.dc.id, path, params);
-    // if (adResponse.status !== 200) {
-    //   const errMsg = (await adResponse.json())['datagrok-error'];
-    //   throw new Error(errMsg);
-    // }
-    //const adRes: Forms.runRes = (await adResponse.json()) as Forms.runRes;
-    const adRes: Forms.runRes = JSON.parse(adResStr) as Forms.runRes;
+    const adRes = (await this.fetchAndCheck(path, params)) as Forms.dockLigandRes;
 
     // const modelList: string[] = wu(adRes.poses.matchAll(/MODEL.*?ENDMDL/gs/* lazy, not greedy */))
     //   .map((ma) => ma[0]).toArray();
@@ -148,11 +181,78 @@ export class AutoDockService implements IAutoDockService {
     // const posesDf = DG.DataFrame.fromColumns([posesCol]);
 
     const posesDf: DG.DataFrame = this.ph.parsePdbqt(adRes.poses, poseColName);
+    return posesDf;
+  }
 
-    const res: AutoDockRunResult = {
-      posesDf: posesDf,
+  async dockLigandColumn(receptor: BiostructureData, ligandCol: DG.Column<string>,
+    autodockGpf: string, poseCount: number = 30, poseColName: string = 'poses', debug: boolean = false
+  ): Promise<DG.DataFrame> {
+    if (receptor.binary || receptor.ext !== 'pdb')
+      throw new Error(`Unsupported receptor ext '${receptor.ext}' or binary, must be 'pdb' string.`);
+
+    let ligandPdbCol: DG.Column<string>;
+    switch (ligandCol.semType) {
+      case DG.SEMTYPE.MOLECULE: {
+        const uh = MoleculeUnitsHandler.getOrCreate(ligandCol);
+        ligandPdbCol = await uh.getAsPdb(this.ph);
+        break;
+      }
+      case DG.SEMTYPE.MOLECULE3D: {
+        const uh = Molecule3DUnitsHandler.getOrCreate(ligandCol);
+        ligandPdbCol = await uh.getAsPdb(this.ph);
+        break;
+      }
+      default:
+        throw new Error(`Unsupported ligand column semantic type, ` +
+          `must be '${DG.SEMTYPE.MOLECULE}' or '${DG.SEMTYPE.MOLECULE3D}'.`);
+    }
+
+    const form: Forms.dockLigandList = {
+      receptor: receptor.data as string,
+      receptor_format: receptor.ext,
+      ligand: ligandPdbCol.toList() as string[],
+      ligand_format: 'pdb',
+      autodock_gpf: autodockGpf,
+      pose_count: poseCount,
     };
-    return res;
+
+    const path = '/autodock/dock_ligand_list';
+    const params: RequestInit = {
+      method: 'POST',
+      headers: {'Content-Type': 'application/json'},
+      body: JSON.stringify(form),
+    };
+    const adRes = await this.fetchAndCheck(path, params) as Forms.dockLigandListRes;
+
+    let posesAllDf: DG.DataFrame | undefined = undefined;
+    for (const [ligandIdx, ligandPoses] of Object.entries(adRes.ligand_results)) {
+      const posesDf = this.ph.parsePdbqt(ligandPoses, poseColName);
+      (posesDf.columns.addNewInt('ligandIdx')).init((rowI) => ligandIdx);
+
+      if (posesAllDf === undefined) {
+        posesAllDf = posesDf;
+      } else {
+        posesAllDf.append(posesAllDf, true);
+      }
+    }
+
+    return posesAllDf!;
+  }
+
+  private async fetchAndCheck(path: string, params: RequestInit): Promise<any> {
+    const adResponse: Response = await grok.dapi.docker.dockerContainers.fetchProxy(this.dc.id, path, params);
+    if (adResponse.status !== 200) {
+      const errMsg = adResponse.statusText;
+      // const errMsg = (await adResponse.json())['datagrok-error'];
+      throw new Error(errMsg);
+    }
+    const adRes = (await adResponse.json()) as Forms.dockLigandRes;
+    if ('datagrok-error' in adRes) {
+      const errVal = adRes['datagrok-error'];
+      const errMsg = errVal ? errVal.toString() : 'Unknown error';
+      throw new Error(errMsg);
+    }
+    return adRes;
   }
 
   static async getSvc(): Promise<IAutoDockService> {
@@ -163,8 +263,8 @@ export class AutoDockService implements IAutoDockService {
 }
 
 export async function _runAutodock(
-  receptor: DG.FileInfo, ligand: DG.FileInfo, x: number, y: number, z: number
-): Promise<AutoDockRunResult | null> {
+  receptor: DG.FileInfo, ligand: DG.FileInfo, npts: GridSize
+): Promise<DG.DataFrame | null> {
   const svc: IAutoDockService = new AutoDockService();
   if (!svc.ready) {
     grok.shell.warning('Autodock container not started yet.');
@@ -172,11 +272,13 @@ export async function _runAutodock(
   }
 
   const receptorStr = await receptor.readAsString();
-  const ligandData: BiostructureData = {binary: true, ext: ligand.extension, data: await ligand.readAsBytes()};
-  return await svc.run(receptorStr, ligandData, new GridSize(x, y, z));
+  const receptorData: BiostructureData = {binary: false, ext: receptor.extension, data: await receptor.readAsString()};
+  const ligandData: BiostructureData = {binary: false, ext: ligand.extension, data: await ligand.readAsString()};
+  const autodockGpf = buildDefaultAutodockGpf(receptor.fileName, npts);
+  return await svc.dockLigand(receptorData, ligandData, autodockGpf);
 }
 
-export async function _runAutodock2(molCol: DG.Column<string>, receptorPdb: string): Promise<void> {
+export async function _runAutodock2(molCol: DG.Column<string>, receptor: BiostructureData): Promise<void> {
   // const receptorPdb: string = await receptorFi.readAsString();
 
   let resDf: DG.DataFrame | undefined = undefined;
@@ -190,14 +292,14 @@ export async function _runAutodock2(molCol: DG.Column<string>, receptorPdb: stri
       const ligandMol = molCol.get(lRowI);
       const ligandPdb = await ph.molToPdb(ligandMol!);
       const ligandData: BiostructureData = {binary: false, data: ligandPdb, ext: 'pdb'};
+      const autodockGpf = buildDefaultAutodockGpf(receptor.options!.name!, new GridSize(40, 40, 40));
 
-      const adRes = await adSvc.run(
-        receptorPdb, ligandData, new GridSize(40, 40, 40), 10);
+      const posesDf = await adSvc.dockLigand(receptor, ligandData, autodockGpf, 10);
 
       if (resDf === undefined)
-        resDf = adRes.posesDf.clone(DG.BitSet.create(adRes.posesDf.rowCount, (_i) => false));
+        resDf = posesDf.clone(DG.BitSet.create(posesDf.rowCount, (_i) => false));
 
-      resDf!.append(adRes.posesDf, true);
+      resDf!.append(posesDf, true);
     } finally {
       const t2 = window.performance.now();
       _package.logger.debug('_runAutodock2(), ' + `ligand: ${lRowI}, ` + `ET: ${t2 - t2} ms, `);