+++ packages/Docking/CHANGELOG.md
@@ -8,3 +8,6 @@
 
 ### Bug fixes
 
+* Fix AutoDock for pose_count param
+* Fix AutoDock convert, add tests
+* Fix AutoDock add clinfo test